@oscarbenjamin 2024-01-10
By merging the other branch in you have brought in the old commit that still has the old email address in its metadata:
https://github.com/sympy/sympy/commit/213d5d9d93dfe05e0dad458f095b3ac13436d7fd.patch