@oscarbenjamin 2024-01-10
> Significantly changed benchmark results (PR vs master)
> 
> ```diff
> | Change   | Before [02eeb233]    | After [2446154f]    |   Ratio | Benchmark (Parameter)                         |
> |----------|----------------------|---------------------|---------|-----------------------------------------------|
> | -        | 28.5±0.2ms           | 15.6±0.2ms          |    0.55 | solve.TimeSparseSystem.time_linsolve_Aaug(20) |
> | -        | 54.4±0.6ms           | 25.1±0.09ms         |    0.46 | solve.TimeSparseSystem.time_linsolve_Aaug(30) |
> | -        | 28.3±0.2ms           | 15.6±0.1ms          |    0.55 | solve.TimeSparseSystem.time_linsolve_Ab(20)   |
> | -        | 54.1±0.2ms           | 24.7±0.1ms          |    0.46 | solve.TimeSparseSystem.time_linsolve_Ab(30)   |
> ```

The benchmarks indicate some speed up from this change. My guess is that speeding up `Matrix.atoms()` makes `linsolve` faster here:
https://github.com/sympy/sympy/blob/02eeb233a4078fb5f97a76bafbb36433691b372b/sympy/solvers/solveset.py#L3129-L3133