@oscarbenjamin 2024-01-10
There are some other places that use `uniquely_names_symbol` in matrices:
```console
$ git grep uniquely sympy/matrices/
sympy/matrices/determinant.py:from sympy.core.symbol import uniquely_named_symbol
sympy/matrices/determinant.py:    x = uniquely_named_symbol(x, M, modify=lambda s: '_' + s)
sympy/matrices/expressions/trace.py:from sympy.core.symbol import uniquely_named_symbol
sympy/matrices/expressions/trace.py:        i = uniquely_named_symbol('i', expr)
sympy/matrices/matrixbase.py:from sympy.core.symbol import uniquely_named_symbol
sympy/matrices/matrixbase.py:        x = uniquely_named_symbol('x', self, modify=lambda s: '_' + s)
sympy/matrices/solvers.py:from sympy.core.symbol import Dummy, uniquely_named_symbol, symbols
sympy/matrices/solvers.py:    name = uniquely_named_symbol('tau', aug,
````
It would be worth investigating whether any of those passes a matrix directly. This one looks suspicious for example:
https://github.com/sympy/sympy/blob/93ceab17149ed2d92e614b1916803624f6577d32/sympy/matrices/solvers.py#L593
```python
In [1]: %time ok = eye(1000).solve(ones(1000, 1))
CPU times: user 30.9 s, sys: 11.5 ms, total: 30.9 s
Wall time: 30.9 s

In [2]: %prun -s cumulative eye(1000).solve(ones(1000, 1))
         64440426 function calls (64438424 primitive calls) in 66.081 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000   66.081   66.081 {built-in method builtins.exec}
        1    0.000    0.000   66.080   66.080 <string>:1(<module>)
        1    0.000    0.000   66.003   66.003 matrixbase.py:5073(solve)
        1    0.000    0.000   66.003   66.003 solvers.py:778(_solve)
        1    0.001    0.001   66.003   66.003 matrixbase.py:5064(gauss_jordan_solve)
        1    0.005    0.005   66.002   66.002 solvers.py:437(_gauss_jordan_solve)
        1    0.000    0.000   65.669   65.669 symbol.py:130(uniquely_named_symbol)
```